Add support for all light-modifying operations
* implement the lights API
